refactor(types): address remaining PR #4171 review feedback
- src/atoms.ts: extract ATOMS/NON_ATOMS/isAtom into a small module so
  render-a11y-string no longer drags the 870-line symbols.ts into its
  bundle.  symbols.ts and parseNode.ts re-export / import from atoms.ts
  to preserve existing import paths.
- delimsizing.ts: make MiddleDelimNode.isMiddle non-optional and
  simplify the two conditionals that previously re-checked isMiddle
  after the type guard.
- defineFunction.ts / defineEnvironment.ts / Parser.ts: replace
  FunctionSpec<any> / EnvSpec<any> registry values with
  FunctionSpec<NodeType> / EnvSpec<NodeType>.  HtmlBuilder<any> and
  MathMLBuilder<any> remain because those generics are contravariant
  in NODETYPE, so no single concrete argument can type the
  heterogeneous builder registry; a comment explains why.
- stretchy.ts: inline the data.length === 4 check and drop the
  isSvgData4 helper as suggested.
